requirements.txt
google-genai>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
pydantic>=2.7.0
python-multipart>=0.0.9
pydantic-settings>=2.5.2
tomli>=2.0.1

